Move software keyblob handling to separate compilation unit
Android has introduced quite a few key blob formats that have
to be supported. The decryptions and parsing code was part
of the sofware keymaster context.

To make it easier to reuse this code by new keymaster
implementations that need this backwards combatilibity,
this patch moves this code to a separate compilation
unit.

Test: VtsHalKeymasterV3_0TargetTest
Bug: 67358942
Change-Id: I98b2d281f555b37e6276795c89f45aab4db7f393