Move Key into Operation
The Keymaster implementation creates a Key object and then passes it
to an Operation object, which copies parts of it.  The Key object is
not needed after the Operation has been created, so much of that
copying is unnecessary.  This CL begins changes that by passing an
rvalue reference to the Key to the Operation, and modifying operations
so they move the pieces of the Key that they need out, wherever
possible.

Test:  make (local unit tests), VTS and CTS
Change-Id: I6c9a27d9ee85ccaeed1efb0fcc3ed0f8694c5771