Refactor AES operations to generalize block cipher operations.
In preparation for adding 3DES support, this CL moves the code that
does all the block cipher work from AesOperation to
EvpBlockCipherOperation (and associated classes).  To make it easier
to see what was changed, the block cipher code was left in
aes_operation.{cpp|h}.  The next CL will move it to separate files.

Test: make (local unit tests), CTS & VTS
Change-Id: Ibbf870c351425ea8d990218aa0ae089d0b2ada4b