Add "unlocked device required" API
Add a keymaster parameter for keys that should be inaccessible when
the device screen is locked. "Locked" here is a state where the device
can be used or accessed without any further trust factor such as a
PIN, password, fingerprint, or trusted face or voice.

This parameter is added to the Java keystore interface for key
creation and import, as well as enums specified by and for the native
keystore process.

This is the API only, full functionality will be added in a later
commit.

Test: CTS tests in CtsKeystoreTestCases

Bug: 67752510

Merged-In: I967193e7f3be4f7a4ddea004b9b203fe035a5d44
Change-Id: I967193e7f3be4f7a4ddea004b9b203fe035a5d44
(cherry picked from commit c9c285e340c42b6c356135bef822f8ed54968019)