Restore "Add "Unlocked device required" parameter to keys"
Add a keymaster parameter for keys that should be inaccessible when
the device screen is locked. "Locked" here is a state where the device
can be used or accessed without any further trust factor such as a
PIN, password, fingerprint, or trusted face or voice.

This parameter is added to the Java keystore interface for key
creation and import, as well as enums specified by and for the native
keystore process.

This reverts commit a1ba90df6372d7194fcb2fa59d69f0d53b2141a6.

Test: CTS tests in I8a5affd1eaed176756175158e3057e44934fffed
(cherry picked from commit cff1364ed2d18b23b987188aededbbb422ce147e)

Bug: 67752510

Merged-In: I62d4176f04077af96779cb53c628a8579a7679d1
Change-Id: I62d4176f04077af96779cb53c628a8579a7679d1