Remove fortify level for memset.
In memset_s, we intentionally turn off the optimization for memset,
this would cause the checking code in bits/string3.h incorrectly
triggered no matter what size we give to memset. The dead code
which should be optimized away would not, and the code would always
throw an error. To fix this problem, we disable the checking
while including string.h, so that string3.h would not be included
because the checking is only intended to work for optimizatized
code.

Tests: Compile the code using gcc 4.8 and define __FORTIFY_SOURCE
to 2, no error is thrown. Note that gcc after 5 would not throw
an error, instead, compiler would give an warning.
Bug: None