Switching keystore to C++17
This header declares nothrow new operators for the use in restricted
(limited stl support) environments, e.g., Trusty. The implementations
are weak so that they can get overridden by the linker in full
environments with full stl support. The software keymaster device gets
used by keystore, which uses this new declaration. As keystore switches
to C++17 the noexcept declaration becomes part of the function type
and thus the compiler must treat the redeclaration without the
keyword as an error. This weak redeclaration of _NOEXCEPT adds the
noexcept keyword as required.

Test: Builds
Bug: 111443219
Change-Id: Ica3c3f0b4d938168d4d671b7cfd2b3881d0cf48d